+++ src/Bitpay/Item.php
@@ -88,6 +88,8 @@ class Item implements ItemInterface
 
     /**
      * @inheritdoc
+     *
+     * @return float
      */
     public function getPrice()
     {
@@ -95,21 +97,17 @@ class Item implements ItemInterface
     }
 
     /**
-     * @param float $price
+     * @param mixed $price A float, integer, or en_US formatted numeric string
      *
      * @return ItemInterface
      */
     public function setPrice($price)
     {
-        // If we get float, then convert it to correct string
-        if(!is_string($price) AND is_float($price)) {
-            $price = number_format($price, 2, '.', '');
+        if (is_string($price)) {
+            $this->checkPriceFormat($price);
         }
-        // Check a format
-        if (1 !== preg_match('/^[0-9]+(?:\.[0-9]{1,2})?$/', $price)) {
-            throw new \Bitpay\Client\ArgumentException("Price must be formatted as a float");
-        }
-        $this->price = $price;
+
+        $this->price = (float)$price;
 
         return $this;
     }
@@ -149,8 +147,22 @@ class Item implements ItemInterface
      */
     public function setPhysical($physical)
     {
-        $this->physical = (boolean) $physical;
+        $this->physical = (boolean)$physical;
 
         return $this;
+    }
+
+    /**
+     * Checks the new price to include BTC
+     * values with more than 6 decimals.
+     *
+     * @param string $price The price value to check
+     * @throws \Exception
+     */
+    protected function checkPriceFormat($price)
+    {
+        if (preg_match('/^[0-9]+?[\.,][0-9]{1,6}?$/', $price) !== 1) {
+            throw new \Bitpay\Client\ArgumentException("Price must be formatted as a float");
+        }
     }
 }