+++ tests/Bitpay/ItemTest.php
@@ -58,15 +58,103 @@ class ItemTest extends \PHPUnit_Framework_TestCase
      */
     public function testSetPrice()
     {
+        setlocale(LC_NUMERIC, 'en_US');
+        // Accepts floats
         $this->item->setPrice(9.99);
+        $this->assertSame(9.99, $this->item->getPrice());
+
+        // Accepts integers
+        $this->item->setPrice(9);
+        $this->assertSame(9.0, $this->item->getPrice());
+
+        // Accepts standard float string
+        $this->item->setPrice("9.99");
         $this->assertNotNull($this->item->getPrice());
         $this->assertSame(9.99, $this->item->getPrice());
         $this->assertInternalType('float', $this->item->getPrice());
 
-        $this->item->setPrice("9.99");
+        $this->item->setPrice("0.999999");
         $this->assertNotNull($this->item->getPrice());
-        $this->assertSame("9.99", $this->item->getPrice());
-        $this->assertInternalType('string', $this->item->getPrice());
+        $this->assertSame(.999999, $this->item->getPrice());
+        $this->assertInternalType('float', $this->item->getPrice());
+
+        // if the argument is a string
+        // and the string is not "floatlike"
+        // throw an error
+    }
+
+    /**
+     *  @expectedException \Bitpay\Client\ArgumentException
+     */
+    public function testBadStringPrice()
+    {
+        // Accepts standard integer string
+        $this->item->setPrice("456");
+        $this->assertSame(456.0, $this->item->getPrice());
+
+        // Accepts float string without leading integer
+        $this->item->setPrice(".99");
+        $this->assertSame(0.99, $this->item->getPrice());
+        $this->item->setPrice(".0");
+        $this->assertSame(0.0, $this->item->getPrice());
+
+        // Accepts float string without decimal numbers
+        $this->item->setPrice("9.");
+        $this->assertSame(9.0, $this->item->getPrice());
+
+        // Arbitrary precision
+        $this->item->setPrice("9.4329082");
+        $this->assertSame(9.4329082, $this->item->getPrice());
+
+        // Try with different locale
+        setlocale(LC_NUMERIC, 'it_IT');
+
+        // Accepts floats
+        $this->item->setPrice(9.99);
+        $this->assertSame(9.99, $this->item->getPrice());
+
+        // Accepts integers
+        $this->item->setPrice(9);
+        $this->assertSame(9.0, $this->item->getPrice());
+
+        // Accepts standard float string
+        $this->item->setPrice("9.99");
+        $this->assertSame(9.99, $this->item->getPrice());
+
+        // Accepts standard integer string
+        $this->item->setPrice("456");
+        $this->assertSame(456.0, $this->item->getPrice());
+
+        // Accepts float string without leading integer
+        $this->item->setPrice(".99");
+        $this->assertSame(0.99, $this->item->getPrice());
+        $this->item->setPrice(".0");
+        $this->assertSame(0.0, $this->item->getPrice());
+
+        // Accepts float string without decimal numbers
+        $this->item->setPrice("9.");
+        $this->assertSame(9.0, $this->item->getPrice());
+
+        // Arbitrary precision
+        $this->item->setPrice("9.4329082");
+        $this->assertSame(9.4329082, $this->item->getPrice());
+        setlocale(LC_NUMERIC, 'en_US');
+    }
+
+    /**
+     * @expectedException        \Bitpay\Client\ArgumentException
+     */
+    public function testSetPriceExceptionNoNumber()
+    {
+        $this->item->setPrice(".");
+    }
+
+    /**
+     * @expectedException        \Bitpay\Client\ArgumentException
+     */
+    public function testSetPriceExceptionDoubleDecimal()
+    {
+        $this->item->setPrice("6..5");
     }
 
     public function testGetQuantity()